+++ tests/testthat/test-helpers.R
@@ -5,11 +5,11 @@ mvfile <- system.file(file.path("extdata/tab", file), package="vapour")
 gfile <- system.file(file.path("extdata", "sst_c.gpkg"), package="vapour")
 osmfile <- system.file(file.path("extdata/osm", "osm-ways.osm"), package="vapour")
 ##dput(vapour_geom_summary(mvfile, sql = sprintf("SELECT FID FROM %s WHERE FID = 15", vapour_layer_names(mvfile)[1L])))
-sql_bench <- list(FID = 15L, valid_geometry = TRUE, xmin = 455980.34, xmax = 463978.72,
+sql_bench <- list(FID = 15L, valid_geometry = TRUE, type = 3L, xmin = 455980.34, xmax = 463978.72,
                   ymin = 5392869.75, ymax = 5399258.91)
 
 test_that("helpers work", {
-  vapour_geom_summary(mvfile) %>% expect_named(c("FID", "valid_geometry", "xmin", "xmax", "ymin", "ymax"))
+  vapour_geom_summary(mvfile) %>% expect_named(c("FID", "valid_geometry", "type", "xmin", "xmax", "ymin", "ymax"))
 
   vapour_geom_summary(mvfile, sql = sprintf("SELECT FID FROM %s WHERE FID = 15", vapour_layer_names(mvfile)[1L])) %>%
     expect_equal(sql_bench)